create metis:conversation:shared
