+++ feature/metis-test/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metistest/MetisServiceStub.kt
@@ -1,7 +1,6 @@
 package de.tum.informatics.www1.artemis.native_app.feature.metistest
 
 import de.tum.informatics.www1.artemis.native_app.core.data.NetworkResponse
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.model.LinkPreview
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.network.MetisService
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.MetisContext
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.StandalonePost
@@ -33,7 +32,7 @@ class MetisServiceStub(
         url: String,
         serverUrl: String,
         authToken: String
-    ): NetworkResponse<LinkPreview?> {
+    ): NetworkResponse<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.model.LinkPreview?> {
         return NetworkResponse.Response(null)
     }
 }
