+++ feature/metis/conversation/build.gradle.kts
@@ -22,7 +22,8 @@ dependencies {
     testImplementation(project(":core:ui-test"))
 
     implementation(project(":feature:metis:shared"))
-    implementation(project(":feature:metis:conversation:saved-posts"))
+    api(project(":feature:metis:conversation:shared"))
+    api(project(":feature:metis:conversation:saved-posts"))
     testImplementation(project(":feature:metis-test"))
 
     implementation(libs.androidx.paging.runtime)
