+++ feature/metis/conversation/saved-posts/build.gradle.kts
@@ -1,7 +1,6 @@
 plugins {
     id("artemis.android.feature")
     id("artemis.android.library.compose")
-    id("kotlin-parcelize")
 }
 
 android {
@@ -12,4 +11,5 @@ dependencies {
     implementation(project(":core:device"))
 
     implementation(project(":feature:metis:shared"))
+    implementation(project(":feature:metis:conversation:shared"))
 }
