+++ feature/metis/conversation/shared/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/shared/service/CreatePostStatus.kt
@@ -1,0 +1,7 @@
+package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service
+
+enum class CreatePostStatus {
+    PENDING,
+    FAILED,
+    FINISHED
+}
