+++ feature/metis/conversation/shared/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/shared/service/model/LinkPreview.kt
@@ -1,4 +1,4 @@
-package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.model
+package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.model
 
 import kotlinx.serialization.Serializable
