+++ feature/metis/conversation/shared/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/shared/ui/BottomSheetActionButton.kt
@@ -1,0 +1,36 @@
+package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.ui
+
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun BottomSheetActionButton(
+    modifier: Modifier,
+    icon: ImageVector,
+    text: String,
+    onClick: () -> Unit
+) {
+    TextButton(
+        modifier = modifier,
+        onClick = onClick
+    ) {
+        Icon(
+            imageVector = icon,
+            contentDescription = null
+        )
+
+        Spacer(modifier = Modifier.width(16.dp))
+
+        Text(
+            modifier = Modifier.weight(1f),
+            text = text
+        )
+    }
+}
