+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/ConversationViewModel.kt
@@ -30,18 +30,17 @@ import de.tum.informatics.www1.artemis.native_app.feature.faq.repository.FaqRepo
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.R
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.saved_posts.service.SavedPostService
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.CreatePostService
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.MetisModificationFailure
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.asMetisModificationFailure
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.model.FileValidationConstants.isImage
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.model.Link
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.model.LinkPreview
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.network.MetisModificationService
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.network.MetisService
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.storage.MetisStorageService
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.storage.ReplyTextStorageService
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.asMetisModificationFailure
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.model.Link
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.model.LinkPreview
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.ui.post.util.LinkPreviewUtil
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.chatlist.ConversationChatListUseCase
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.post.post_actions.PostActionFlags
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.post.util.LinkPreviewUtil
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.InitialReplyTextProvider
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.AutoCompletionUseCase
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.thread.ConversationThreadUseCase
@@ -311,17 +310,17 @@ internal open class ConversationViewModel(
         post: IBasePost,
         emojiId: String,
         create: Boolean
-    ): Deferred<MetisModificationFailure?> {
+    ): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         return viewModelScope.async(coroutineContext) {
             if (create) {
                 createReactionImpl(
                     emojiId,
                     post.getAsAffectedPost()
-                        ?: return@async MetisModificationFailure.DELETE_REACTION
+                        ?: return@async de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.DELETE_REACTION
                 )
             } else {
                 val clientId = clientId.value.orNull()
-                    ?: return@async MetisModificationFailure.DELETE_REACTION
+                    ?: return@async de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.DELETE_REACTION
 
                 val exitingReactionId = post
                     .reactions
@@ -329,7 +328,7 @@ internal open class ConversationViewModel(
                     .filter { it.emojiId == emojiId }
                     .firstOrNull { it.creatorId == clientId }
                     ?.id
-                    ?: return@async MetisModificationFailure.DELETE_REACTION
+                    ?: return@async de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.DELETE_REACTION
 
                 deleteReactionImpl(exitingReactionId)
             }
@@ -339,7 +338,7 @@ internal open class ConversationViewModel(
     private suspend fun createReactionImpl(
         emojiId: String,
         post: MetisModificationService.AffectedPost
-    ): MetisModificationFailure? {
+    ): de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure? {
         val networkResponse: NetworkResponse<Reaction> = metisModificationService.createReaction(
             context = metisContext,
             post = post,
@@ -348,11 +347,11 @@ internal open class ConversationViewModel(
             authToken = accountService.authToken.first()
         )
 
-        return networkResponse.bind<MetisModificationFailure?> { null }
-            .or(MetisModificationFailure.CREATE_REACTION)
+        return networkResponse.bind<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> { null }
+            .or(de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.CREATE_REACTION)
     }
 
-    private suspend fun deleteReactionImpl(reactionId: Long): MetisModificationFailure? {
+    private suspend fun deleteReactionImpl(reactionId: Long): de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure? {
         val success = metisModificationService.deleteReaction(
             context = metisContext,
             reactionId = reactionId,
@@ -360,7 +359,7 @@ internal open class ConversationViewModel(
             authToken = accountService.authToken.first()
         ).or(false)
 
-        return if (success) null else MetisModificationFailure.DELETE_REACTION
+        return if (success) null else de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.DELETE_REACTION
     }
 
     /**
@@ -370,10 +369,10 @@ internal open class ConversationViewModel(
     fun toggleResolvePost(
         parentPost: IStandalonePost,
         post: AnswerPostPojo
-    ): Deferred<MetisModificationFailure?> {
+    ): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         return viewModelScope.async(coroutineContext) {
             val conversation =
-                loadConversation() ?: return@async MetisModificationFailure.UPDATE_POST
+                loadConversation() ?: return@async de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.UPDATE_POST
 
             val resolved = !post.resolvesPost
             val serializedParentPost = StandalonePost(parentPost, conversation)
@@ -385,14 +384,14 @@ internal open class ConversationViewModel(
                 serverUrl = serverConfigurationService.serverUrl.first(),
                 authToken = accountService.authToken.first()
             )
-                .asMetisModificationFailure(MetisModificationFailure.UPDATE_POST)
+                .asMetisModificationFailure(de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.UPDATE_POST)
         }
     }
 
-    fun togglePinPost(post: IStandalonePost): Deferred<MetisModificationFailure?> {
+    fun togglePinPost(post: IStandalonePost): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         return viewModelScope.async(coroutineContext) {
             val conversation =
-                loadConversation() ?: return@async MetisModificationFailure.UPDATE_POST
+                loadConversation() ?: return@async de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.UPDATE_POST
 
             val newDisplayPriority = if (post.displayPriority == DisplayPriority.PINNED) {
                 DisplayPriority.NONE
@@ -416,11 +415,11 @@ internal open class ConversationViewModel(
                 serverUrl = serverConfigurationService.serverUrl.first(),
                 authToken = accountService.authToken.first(),
             )
-                .asMetisModificationFailure(MetisModificationFailure.UPDATE_POST)
+                .asMetisModificationFailure(de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.UPDATE_POST)
         }
     }
 
-    fun toggleSavePost(post: IBasePost): Deferred<MetisModificationFailure?> {
+    fun toggleSavePost(post: IBasePost): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         // TODO: this is a quick fix to prevent multiple save requests.
         //      https://github.com/ls1intum/artemis-android/issues/307
         if (currentlySavingPost) return CompletableDeferred(null)
@@ -443,12 +442,12 @@ internal open class ConversationViewModel(
 
             currentlySavingPost = false
             response.bind { requestReload() }   // Currently changing save status does not trigger a websocket update
-                .asMetisModificationFailure(MetisModificationFailure.UPDATE_POST)
+                .asMetisModificationFailure(de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.UPDATE_POST)
 
         }
     }
 
-    fun deletePost(post: IBasePost): Deferred<MetisModificationFailure?> {
+    fun deletePost(post: IBasePost): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         isMarkedAsDeleteList.add(post)
         deleteJobs[post]?.cancel()
 
@@ -459,13 +458,13 @@ internal open class ConversationViewModel(
                 if (isMarkedAsDeleteList.contains(post)) {
                     metisModificationService.deletePost(
                         metisContext,
-                        post.getAsAffectedPost() ?: return@async MetisModificationFailure.DELETE_POST,
+                        post.getAsAffectedPost() ?: return@async de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.DELETE_POST,
                         serverConfigurationService.serverUrl.first(),
                         accountService.authToken.first()
                     )
-                        .bind { if (it) null else MetisModificationFailure.DELETE_POST }
-                        .or(MetisModificationFailure.DELETE_POST)
-                        .also { if (it != MetisModificationFailure.DELETE_POST && post is IStandalonePost) onCloseThread?.invoke() }
+                        .bind { if (it) null else de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.DELETE_POST }
+                        .or(de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.DELETE_POST)
+                        .also { if (it != de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.DELETE_POST && post is IStandalonePost) onCloseThread?.invoke() }
                 } else {
                     null
                 }
@@ -485,10 +484,10 @@ internal open class ConversationViewModel(
         isMarkedAsDeleteList.remove(post)
     }
 
-    fun editPost(post: IStandalonePost, newText: String): Deferred<MetisModificationFailure?> {
+    fun editPost(post: IStandalonePost, newText: String): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         return viewModelScope.async(coroutineContext) {
             val conversation =
-                loadConversation() ?: return@async MetisModificationFailure.UPDATE_POST
+                loadConversation() ?: return@async de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.UPDATE_POST
 
             val newPost = when (post) {
                 is StandalonePost -> post.copy(content = newText)
@@ -506,7 +505,7 @@ internal open class ConversationViewModel(
                 serverUrl = serverConfigurationService.serverUrl.first(),
                 authToken = accountService.authToken.first()
             )
-                .asMetisModificationFailure(MetisModificationFailure.UPDATE_POST)
+                .asMetisModificationFailure(de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.UPDATE_POST)
         }
     }
 
@@ -514,10 +513,10 @@ internal open class ConversationViewModel(
         parentPost: IStandalonePost,
         post: AnswerPostPojo,
         newText: String
-    ): Deferred<MetisModificationFailure?> {
+    ): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         return viewModelScope.async(coroutineContext) {
             val conversation =
-                loadConversation() ?: return@async MetisModificationFailure.UPDATE_POST
+                loadConversation() ?: return@async de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.UPDATE_POST
 
             val serializedParentPost = StandalonePost(parentPost, conversation)
             val newPost = AnswerPost(post, serializedParentPost).copy(content = newText)
@@ -528,7 +527,7 @@ internal open class ConversationViewModel(
                 serverUrl = serverConfigurationService.serverUrl.first(),
                 authToken = accountService.authToken.first()
             )
-                .asMetisModificationFailure(MetisModificationFailure.UPDATE_POST)
+                .asMetisModificationFailure(de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.UPDATE_POST)
         }
     }
 
@@ -536,7 +535,7 @@ internal open class ConversationViewModel(
         linkPreview: LinkPreview,
         post: IBasePost,
         parentPost: IStandalonePost?
-    ): Deferred<MetisModificationFailure?> {
+    ): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         val newContent = LinkPreviewUtil.removeLinkPreview(post.content.orEmpty(), linkPreview.url)
         return when (post) {
             is IStandalonePost -> editPost(post, newContent)
@@ -672,7 +671,7 @@ internal open class ConversationViewModel(
             null -> null
         }
 
-    fun createPost(): Deferred<MetisModificationFailure?> {
+    fun createPost(): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         createPostService.createPost(courseId, conversationId, newMessageText.value.text)
 
         return CompletableDeferred(value = null)
@@ -701,10 +700,10 @@ internal open class ConversationViewModel(
         }
     }
 
-    fun createReply(): Deferred<MetisModificationFailure?> {
+    fun createReply(): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> {
         return viewModelScope.async(coroutineContext) {
             val serverSideParentPostId =
-                getPostId() ?: return@async MetisModificationFailure.CREATE_POST
+                getPostId() ?: return@async de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure.CREATE_POST
 
             createPostService.createAnswerPost(
                 courseId,
