+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/post/PostItem.kt
@@ -2,9 +2,6 @@ package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.SizeTransform
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.slideInVertically
@@ -18,10 +15,8 @@ import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -29,8 +24,6 @@ import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.sizeIn
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
@@ -42,11 +35,9 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableLongStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
@@ -54,8 +45,6 @@ import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.drawWithCache
-import androidx.compose.ui.graphics.BlendMode
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalDensity
@@ -63,18 +52,15 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.pluralStringResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import de.tum.informatics.www1.artemis.native_app.core.ui.Spacings
-import de.tum.informatics.www1.artemis.native_app.core.ui.date.DateFormats
-import de.tum.informatics.www1.artemis.native_app.core.ui.date.format
-import de.tum.informatics.www1.artemis.native_app.core.ui.date.getRelativeTime
-import de.tum.informatics.www1.artemis.native_app.core.ui.markdown.MarkdownText
 import de.tum.informatics.www1.artemis.native_app.core.ui.material.colors.PostColors
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.R
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.CreatePostService
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.model.LinkPreview
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.CreatePostStatus
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.model.LinkPreview
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.ui.post.PostItemMainContent
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.getUnicodeForEmojiId
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.post.post_actions.EmojiDialog
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.post.post_actions.EmojiSelection
@@ -85,16 +71,7 @@ import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.d
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.IBasePost
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.IReaction
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.IStandalonePost
-import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.UserRole
-import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.ui.UserRoleBadge
-import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.ui.profile_picture.ProfilePictureWithDialog
-import io.github.fornewid.placeholder.material3.placeholder
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-import kotlinx.datetime.Clock
-import kotlinx.datetime.Instant
 import org.koin.compose.koinInject
-import java.time.Duration
 
 sealed class PostItemViewType {
 
@@ -106,8 +83,6 @@ sealed class PostItemViewType {
 
     data object ThreadAnswerItem : PostItemViewType()
 }
-
-private const val PlaceholderContent = "WWWWWWW"
 
 /**
  * Displays a post item or a placeholder for it.
@@ -144,11 +119,11 @@ internal fun PostItem(
     // Retrieve post status
     val clientPostId = post?.clientPostId
     val postStatus = when {
-        post == null || post.serverPostId != null || clientPostId == null -> CreatePostService.Status.FINISHED
+        post == null || post.serverPostId != null || clientPostId == null -> CreatePostStatus.FINISHED
         else -> {
             val createPostService: CreatePostService = koinInject()
             createPostService.observeCreatePostWorkStatus(clientPostId)
-                .collectAsState(initial = CreatePostService.Status.PENDING)
+                .collectAsState(initial = CreatePostStatus.PENDING)
                 .value
         }
     }
@@ -156,7 +131,7 @@ internal fun PostItem(
     PostItemMainContent(
         modifier = modifier
             .let {
-                if (postStatus == CreatePostService.Status.FAILED) {
+                if (postStatus == CreatePostStatus.FAILED) {
                     it
                         .background(color = MaterialTheme.colorScheme.errorContainer)
                         .clickable(onClick = onRequestRetrySend)
@@ -262,159 +237,6 @@ internal fun PostItem(
     )
 }
 
-@Composable
-fun PostItemMainContent(
-    modifier: Modifier = Modifier,
-    post: IBasePost?,
-    isExpanded: Boolean = true,
-    isPlaceholder: Boolean = false,
-    isDeleting: Boolean = false,
-    isAuthor: Boolean = false,
-    postStatus: CreatePostService.Status = CreatePostService.Status.FINISHED,
-    displayHeader: Boolean = true,
-    linkPreviews: List<LinkPreview> = emptyList(),
-    onRemoveLinkPreview: (LinkPreview) -> Unit = {},
-    onClick: () -> Unit,
-    onLongClick: () -> Unit,
-    onUndoDelete: () -> Unit = {},
-    leadingContent: @Composable ColumnScope.() -> Unit = {},
-    trailingContent: @Composable ColumnScope.() -> Unit = {}
-) {
-    Column(
-        modifier = modifier
-    ) {
-        leadingContent()
-
-        PostHeadline(
-            modifier = Modifier.fillMaxWidth(),
-            postStatus = postStatus,
-            authorRole = post?.authorRole,
-            authorName = post?.authorName,
-            authorId = post?.authorId ?: -1,
-            authorImageUrl = post?.authorImageUrl,
-            creationDate = post?.creationDate,
-            expanded = isExpanded,
-            isAnswerPost = post is IAnswerPost,
-            displayHeader = displayHeader,
-            isDeleting = isDeleting
-        ) {
-            Column(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                if (isDeleting) {
-                    UndoDeleteHeader(
-                        modifier = Modifier.fillMaxWidth(),
-                        onClick = onUndoDelete
-                    )
-                    return@PostHeadline
-                }
-
-                MarkdownText(
-                    markdown = remember(post?.content, isPlaceholder) {
-                        if (isPlaceholder) {
-                            PlaceholderContent
-                        } else post?.content.orEmpty()
-                    },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .placeholder(visible = isPlaceholder),
-                    style = MaterialTheme.typography.bodyMedium,
-                    onClick = onClick,
-                    onLongClick = onLongClick,
-                    color = if (post?.serverPostId == null) PostColors.unsentMessageText else Color.Unspecified
-                )
-
-                val instant = post?.updatedDate
-                if (instant != null) {
-                    Spacer(modifier = Modifier.height(Spacings.Post.innerSpacing))
-
-                    val updateTime = instant.format(DateFormats.EditTimestamp.format)
-                    Text(
-                        text = stringResource(id = R.string.post_edited_hint, updateTime),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = PostColors.editedHintText
-                    )
-                }
-
-                if (linkPreviews.isNotEmpty()) {
-                    Spacer(modifier = Modifier.height(Spacings.Post.innerSpacing))
-
-                    LinkPreviewColumn(
-                        modifier = Modifier.fillMaxWidth(),
-                        linkPreviews = linkPreviews,
-                        isAuthor = isAuthor,
-                        onRemoveLinkPreview = onRemoveLinkPreview
-                    )
-                }
-
-                trailingContent()
-            }
-        }
-    }
-}
-
-@Composable
-private fun PostHeadline(
-    modifier: Modifier,
-    authorRole: UserRole?,
-    authorName: String?,
-    authorId: Long,
-    authorImageUrl: String?,
-    creationDate: Instant?,
-    postStatus: CreatePostService.Status,
-    expanded: Boolean = false,
-    isAnswerPost: Boolean,
-    displayHeader: Boolean = true,
-    isDeleting: Boolean,
-    content: @Composable () -> Unit
-) {
-    val doDisplayHeader = displayHeader || postStatus == CreatePostService.Status.FAILED
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(Spacings.Post.innerSpacing)
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(Spacings.Post.innerSpacing)
-        ) {
-            if (!doDisplayHeader) {
-                return@Row
-            }
-
-            HeadlineProfilePicture(
-                userId = authorId,
-                userName = authorName.orEmpty(),
-                imageUrl = authorImageUrl,
-                userRole = authorRole,
-                isGrayscale = isDeleting
-            )
-
-            if (postStatus == CreatePostService.Status.FAILED) {
-                Text(
-                    text = stringResource(id = R.string.post_sending_failed),
-                    style = MaterialTheme.typography.labelMedium,
-                    color = MaterialTheme.colorScheme.error
-                )
-            }
-
-            HeadlineAuthorInfo(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .wrapContentHeight(),
-                authorName = authorName,
-                authorRole = authorRole,
-                creationDate = creationDate,
-                expanded = expanded,
-                isAnswerPost = isAnswerPost,
-                isGrayscale = isDeleting
-            )
-        }
-
-        content()
-    }
-}
 
 @Composable
 private fun IconLabel(
@@ -437,119 +259,6 @@ private fun IconLabel(
         Text(
             text = stringResource(id = resourceString),
             style = MaterialTheme.typography.bodySmall
-        )
-    }
-}
-
-@Composable
-private fun HeadlineAuthorInfo(
-    modifier: Modifier,
-    authorName: String?,
-    authorRole: UserRole?,
-    creationDate: Instant?,
-    expanded: Boolean,
-    isAnswerPost: Boolean,
-    isGrayscale: Boolean
-) {
-    Column(modifier = modifier) {
-        AuthorRoleAndTimeRow(
-            expanded = expanded,
-            authorRole = authorRole,
-            creationDate = creationDate,
-            isAnswerPost = isAnswerPost,
-            isGrayscale = isGrayscale
-        )
-
-        Spacer(modifier = Modifier.weight(1f))
-
-        Text(
-            modifier = Modifier,
-            text = remember(authorName) { authorName ?: "Placeholder" },
-            maxLines = 1,
-            style = MaterialTheme.typography.titleSmall,
-            fontWeight = FontWeight.Bold
-        )
-    }
-}
-
-@Composable
-private fun AuthorRoleAndTimeRow(
-    expanded: Boolean,
-    authorRole: UserRole?,
-    creationDate: Instant?,
-    isAnswerPost: Boolean,
-    isGrayscale: Boolean
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        val relativeTimeTo = remember(creationDate) {
-            creationDate ?: Clock.System.now()
-        }
-
-        val creationDateContent: @Composable () -> Unit = {
-
-            val relativeTime = if (expanded || isAnswerPost) {
-                getRelativeTime(to = relativeTimeTo, showDateAndTime = true)
-            } else {
-                getRelativeTime(to = relativeTimeTo, showDate = false)
-            }
-
-            Text(
-                modifier = Modifier,
-                text = relativeTime.toString(),
-                style = MaterialTheme.typography.bodySmall
-            )
-        }
-
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-
-            /*
-            * remember is needed here to prevent the value from being reset after an update
-            * (the author role is not sent when updating a post)
-            */
-            val initialAuthorRole by remember { mutableStateOf(authorRole) }
-            UserRoleBadge(
-                modifier = Modifier.applyGrayscale(isGrayscale),
-                userRole = initialAuthorRole
-            )
-            Spacer(modifier = Modifier.weight(1f))
-            creationDateContent()
-        }
-    }
-}
-
-@Composable
-private fun HeadlineProfilePicture(
-    userId: Long,
-    userName: String,
-    imageUrl: String?,
-    userRole: UserRole?,
-    displayImage: Boolean = true,
-    isGrayscale: Boolean = false
-) {
-    val fontScale = LocalDensity.current.fontScale
-    val scaledSizeDp = Spacings.Post.postHeadlineHeight * fontScale
-
-    Box(
-        modifier = Modifier
-            .sizeIn(minWidth = scaledSizeDp, minHeight = scaledSizeDp)
-            .applyGrayscale(isGrayscale)
-    ) {
-        if (!displayImage) {
-            return
-        }
-
-        ProfilePictureWithDialog(
-            modifier = Modifier.sizeIn(minWidth = scaledSizeDp, minHeight = scaledSizeDp),
-            userId = userId,
-            userName = userName,
-            userRole = userRole,
-            imageUrl = imageUrl,
         )
     }
 }
@@ -732,87 +441,5 @@ private fun EmojiChip(
 
             AnimatedCounter(reactionCount, selected)
         }
-    }
-}
-
-@Composable
-fun UndoDeleteHeader(
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit,
-    delay: Duration = Duration.ofSeconds(6)
-) {
-    var remainingTime by remember { mutableLongStateOf(delay.seconds) }
-    val animatedProgress = remember { Animatable(0f) }
-
-    LaunchedEffect(Unit) {
-        launch {
-            animatedProgress.animateTo(
-                targetValue = 1f,
-                animationSpec = tween(
-                    durationMillis = delay.toMillis().toInt(),
-                    easing = LinearEasing
-                )
-            )
-        }
-        while (remainingTime > 0) {
-            delay(Duration.ofSeconds(1).toMillis())
-            remainingTime--
-        }
-    }
-
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        Text(
-            modifier = Modifier.weight(1f),
-            text = pluralStringResource(
-                id = R.plurals.post_is_being_deleted,
-                count = remainingTime.toInt(),
-                remainingTime.toInt()
-            ),
-            color = MaterialTheme.colorScheme.secondary,
-            style = MaterialTheme.typography.bodyMedium,
-        )
-
-        Box(
-            modifier = Modifier
-                .width(120.dp)
-                .height(34.dp)
-                .clip(CircleShape)
-                .background(MaterialTheme.colorScheme.primary.copy(alpha = 0.7f))
-                .clickable(onClick = onClick)
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxHeight()
-                    .fillMaxWidth(animatedProgress.value)
-                    .background(MaterialTheme.colorScheme.primary),
-            )
-
-            Text(
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .padding(4.dp),
-                text = stringResource(id = R.string.post_undo_delete),
-                color = MaterialTheme.colorScheme.onPrimary,
-                style = MaterialTheme.typography.bodyMedium,
-            )
-        }
-    }
-}
-
-fun Modifier.applyGrayscale(isGrayscale: Boolean): Modifier {
-    return if (isGrayscale) {
-        this
-            .drawWithCache {
-                onDrawWithContent {
-                    drawContent()
-                    drawRect(Color.Black, blendMode = BlendMode.Saturation)
-                }
-            }
-    } else {
-        this
     }
 }
