+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/ReplyMode.kt
@@ -10,8 +10,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.text.input.TextFieldValue
 import de.tum.informatics.www1.artemis.native_app.core.ui.AwaitDeferredCompletion
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.MetisModificationFailure
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.shared.MetisModificationTaskHandler
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.IBasePost
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.Deferred
@@ -24,19 +22,19 @@ internal sealed class ReplyMode() {
     data class NewMessage(
         override val currentText: State<TextFieldValue>,
         private val onUpdateTextUpstream: (TextFieldValue) -> Unit,
-        private val onCreateNewMessageUpstream: () -> Deferred<MetisModificationFailure?>,
+        private val onCreateNewMessageUpstream: () -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>,
     ) : ReplyMode() {
 
         override fun onUpdate(new: TextFieldValue) {
             onUpdateTextUpstream(new)
         }
 
-        fun onCreateNewMessage(): Deferred<MetisModificationFailure?> = onCreateNewMessageUpstream()
+        fun onCreateNewMessage(): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> = onCreateNewMessageUpstream()
     }
 
     data class EditMessage(
         val post: IBasePost,
-        private val onEditMessage: (String) -> Deferred<MetisModificationFailure?>,
+        private val onEditMessage: (String) -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>,
         val onCancelEditMessage: () -> Unit
     ) : ReplyMode() {
         override val currentText = mutableStateOf(TextFieldValue(post.content ?: ""))
@@ -45,7 +43,7 @@ internal sealed class ReplyMode() {
             currentText.value = new
         }
 
-        fun onEditMessage(): Deferred<MetisModificationFailure?> = onEditMessage(currentText.value.text)
+        fun onEditMessage(): Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> = onEditMessage(currentText.value.text)
     }
 }
 
@@ -54,8 +52,8 @@ private fun <T : IBasePost> rememberReplyMode(
     initialReplyTextProvider: InitialReplyTextProvider,
     editingPost: T?,
     onClearEditingPost: () -> Unit,
-    onCreatePost: () -> Deferred<MetisModificationFailure?>,
-    onEditPost: (T, String) -> Deferred<MetisModificationFailure?>
+    onCreatePost: () -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>,
+    onEditPost: (T, String) -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>
 ): ReplyMode {
     val newMessageText = initialReplyTextProvider.newMessageText.collectAsState(initial = TextFieldValue())
 
@@ -67,7 +65,7 @@ private fun <T : IBasePost> rememberReplyMode(
                 onCreateNewMessageUpstream = onCreatePost
             )
         }
-    var editingPostJob: Deferred<MetisModificationFailure?>? by remember() { mutableStateOf(null) }
+    var editingPostJob: Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>? by remember() { mutableStateOf(null) }
 
     AwaitDeferredCompletion(job = editingPostJob) {
         onClearEditingPost()
@@ -103,13 +101,13 @@ private fun <T : IBasePost> rememberReplyMode(
 @Composable
 internal fun <T : IBasePost> MetisReplyHandler(
     initialReplyTextProvider: InitialReplyTextProvider,
-    onCreatePost: () -> Deferred<MetisModificationFailure?>,
-    onEditPost: (T, String) -> Deferred<MetisModificationFailure?>,
-    onResolvePost: ((T) -> Deferred<MetisModificationFailure?>)?,
-    onPinPost: ((T) -> Deferred<MetisModificationFailure?>)?,
-    onSavePost: ((T) -> Deferred<MetisModificationFailure?>)?,
-    onDeletePost: (T) -> Deferred<MetisModificationFailure?>,
-    onRequestReactWithEmoji: (T, emojiId: String, create: Boolean) -> Deferred<MetisModificationFailure?>,
+    onCreatePost: () -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>,
+    onEditPost: (T, String) -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>,
+    onResolvePost: ((T) -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>)?,
+    onPinPost: ((T) -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>)?,
+    onSavePost: ((T) -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>)?,
+    onDeletePost: (T) -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>,
+    onRequestReactWithEmoji: (T, emojiId: String, create: Boolean) -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>,
     content: @Composable (
         replyMode: ReplyMode,
         onRequestEditPostDelegate: (T) -> Unit,
@@ -118,13 +116,15 @@ internal fun <T : IBasePost> MetisReplyHandler(
         onDeletePostDelegate: (T) -> Unit,
         onPinPostDelegate: (T) -> Unit,
         onSavedPostDelegate: (T) -> Unit,
-        updateFailureStateDelegate: (MetisModificationFailure?) -> Unit
+        updateFailureStateDelegate: (de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?) -> Unit
     ) -> Unit
 ) {
-    var metisModificationTask: Deferred<MetisModificationFailure?>? by remember {
+    var metisModificationTask: Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>? by remember {
         mutableStateOf(null)
     }
-    MetisModificationTaskHandler(metisModificationTask)
+    de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.ui.MetisModificationTaskHandler(
+        metisModificationTask
+    )
 
 
     var editingPost: T? by remember { mutableStateOf(null) }
