+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/ReplyState.kt
@@ -8,7 +8,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.text.input.TextFieldValue
 import de.tum.informatics.www1.artemis.native_app.core.ui.AwaitDeferredCompletion
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.MetisModificationFailure
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.delay
 import kotlin.time.Duration.Companion.seconds
@@ -29,9 +28,9 @@ internal sealed class ReplyState {
 @Composable
 internal fun rememberReplyState(
     replyMode: ReplyMode,
-    updateFailureState: (MetisModificationFailure?) -> Unit
+    updateFailureState: (de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?) -> Unit
 ): ReplyState {
-    var isCreatingReplyJob: Deferred<MetisModificationFailure?>? by remember { mutableStateOf(null) }
+    var isCreatingReplyJob: Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?>? by remember { mutableStateOf(null) }
     var displaySendSuccess by remember { mutableStateOf(false) }
 
     AwaitDeferredCompletion(job = isCreatingReplyJob) { failure ->
