+++ feature/metis/conversation/src/main/res/values/conversation_strings.xml
@@ -27,15 +27,6 @@
     <string name="metis_post_list_error_try_again">Try again</string>
     <string name="metis_post_search_hint">Search for a post</string>
 
-    <!-- Strings related to failures when interacting and modifying metis -->
-    <string name="metis_modification_failure_dialog_title">Failure</string>
-    <string name="metis_modification_failure_dialog_confirm_button">@android:string/ok</string>
-    <string name="metis_modification_failure_dialog_message_create_reaction">Could not create reaction.</string>
-    <string name="metis_modification_failure_dialog_message_delete_reaction">Could not delete reaction.</string>
-    <string name="metis_modification_failure_dialog_message_create_post">Could not create post.</string>
-    <string name="metis_modification_failure_dialog_message_delete_post">Could not delete post.</string>
-    <string name="metis_modification_failure_dialog_message_update_post">Could not update post.</string>
-    <string name="metis_modification_failure_dialog_message_change_saved_post_status">Could change the saved post\'s status.</string>
 
     <string name="standalone_post_loading">Loading post…</string>
     <string name="standalone_post_failure">Something went wrong while loading the post.</string>
