+++ feature/metis/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/ui/SinglePageConversationBody.kt
@@ -197,7 +197,7 @@ internal fun SinglePageConversationBody(
                 is OpenedSavedPosts -> {
                     // TODO: This should potentially be moved into the ConversationScreen. That allows us to still display the ConvOverview on the left.
                     //      https://github.com/ls1intum/artemis-android/issues/288
-                    de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.saved_posts.ui.SavedPostsScreen(
+                    SavedPostsScreen(
                         modifier = modifier,
                         courseId = courseId,
                         savedPostStatus = config.status,
